backend/requirements.txt
fastapi
uvicorn[standard]

psycopg2-binary

pydantic
python-jose[cryptography]
passlib[bcrypt]
